requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
black==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
executing==1.2.0
flake8==6.0.0
idna==3.4
ipdb==0.13.11
ipython==8.10.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyfcm==1.5.4
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==1.7.1
pylint==2.16.2
pytz==2022.7.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.6
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
wrapt==1.14.1
